refactor: moved email template controller to admin
